Introduce Sinatra namespaces
Split up the Users & Organizations controllers into modules and then include
those modules into the correct position in the namespace hierachy.
Generally fix up my horrible Sinatra abuse and be more idiomatic about things.
Add the sinatra-contrib Gem for Sinatra::Namespace
