+++ Gemfile
@@ -4,6 +4,7 @@ gem 'oj'
 gem 'awesome_print'
 gem 'require_all'
 gem 'sinatra'
+gem 'sinatra-contrib'
 gem 'warden'
 gem 'activerecord'
 gem 'sinatra-activerecord'
